fix: :recycle: Refatora arquivo de Permissions
