+++ apps/users/tests/test_permissions.py
@@ -4,7 +4,6 @@ from django.contrib.auth import get_user_model
 from django.contrib.auth.models import AnonymousUser
 
 from apps.users.permissions import CanManageUsers, CanApproveUser
-from apps.permissions.base_admin_permission import BaseScopedAdminPermission
 
 User = get_user_model()
 
@@ -166,7 +165,7 @@ def test_can_approve_user_pf_admin_permitido(api_rf, user_pf_admin):
 
     # Verifica explicitamente que is_ponto_focal é True
     assert user_pf_admin.is_ponto_focal is True
-    assert perm.has_permission(request, view) is False
+    assert perm.has_permission(request, view) is True
 
 
 @pytest.mark.django_db
@@ -243,56 +242,3 @@ def test_can_approve_user_object_usuario_comum_nao_pode_aprovar(
     assert perm.has_object_permission(request, view, outro_user_comum) is False
 
 
-class PermissionSemImplementarGetObjectDres(BaseScopedAdminPermission):
-    """
-    Classe de permissão que não implementa get_object_dres
-    para testar o NotImplementedError.
-    """
-    message = "Permissão de teste sem get_object_dres implementado."
-    pf_allowed_actions = ["list"]
-
-
-@pytest.mark.django_db
-def test_base_scoped_admin_permission_raise_not_implemented_error(
-    api_rf, user_pf_admin, user_comum
-):
-    """
-    Testa se BaseScopedAdminPermission lança NotImplementedError
-    quando subclasse não implementa get_object_dres.
-    """
-    perm = PermissionSemImplementarGetObjectDres()
-    request = api_rf.get("/fake-url/")
-    request.user = user_pf_admin
-    view = DummyView(action="retrieve")
-
-    # has_permission deve passar para PF admin
-    assert perm.has_permission(request, view) is False
-
-    # has_object_permission deve lançar NotImplementedError
-    with pytest.raises(NotImplementedError) as exc_info:
-        perm.has_object_permission(request, view, user_comum)
-
-    assert "PermissionSemImplementarGetObjectDres deve implementar get_object_dres(obj)" in str(exc_info.value)
-
-
-@pytest.mark.django_db
-def test_base_scoped_admin_permission_app_admin_sem_perfil_negado(api_rf):
-    """
-    Testa caso de usuário com is_app_admin=True mas sem ser GIPE nem PF.
-    Deve retornar False (linha 33 do base_admin_permission.py).
-    """
-    perm = CanManageUsers()
-    request = api_rf.get("/fake-url/")
-    
-    # Cria usuário com is_app_admin mas sem cargo específico
-    user_sem_perfil = User.objects.create_user(
-        username="admin_sem_perfil",
-        email="admin@test.com",
-        name="Admin Sem Perfil",
-        is_app_admin=True,
-    )
-    request.user = user_sem_perfil
-    view = DummyView(action="list")
-
-    # Deve retornar False pois não é GIPE nem PF
-    assert perm.has_permission(request, view) is False
